backend/requirements.txt
# Backend requirements
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7  # For PostgreSQL connection
pyjwt>=2.8.0
python-multipart>=0.0.6  # For form parsing in FastAPI
alembic>=1.12.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1
yfinance>=0.2.28
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.4.0

# Development requirements
pytest>=7.4.0
httpx>=0.24.1  # For testing FastAPI
